Correct raycast skipping bug
Sigh. Last fix was done wrong--owing to my use of continue, the horizontal-ray height only got incremented on connecting hits. The increment needs to happen after doing the raycast, before the conditional continues. Despite VerticalCollisions not having any breaks either, moved the increment in that function too for the sake of consistency.